cscope: make it possible to add the new keywords
On 03/12, Oleg Nesterov wrote:
>
> Off-list, because I have no idea where should I send this patch.

Hans, et al, please tell me if you are not interested in this spam
and I will stop it ;)

But now that I looked into cscope sources I think another simple
patch can make it more useful on linux kernel.

The linux kernel uses a lot of annotations which confuse cscope.
Simple example:

	$ cat -n test.c
	     1	void func1(void) __acquires(RCU)
	     2	{
	     3		X = 1;
	     4	}
	     5
	     6	void func2(void) __releases(RCU)
	     7	{
	     8		X = 1;
	     9	}
	$ ./cscope -bu test.c
	$ ./cscope -dL -0 X
	test.c __acquires 3 X = 1;
	test.c __releases 8 X = 1;

With the patch below I can do

	$ ./cscope -bu -K __acquires,__releases test.c

and now

	$ ./cscope -dL -0 X
	test.c func1 3 X = 1;
	test.c func2 8 X = 1;

looks much better to me.

Of course, this is a hack and this way cscope can't report the usage
of __acquires/__releases, but imo this is better than what we have now.

Signed-off-by: Oleg Nesterov <oleg@redhat.com>